change conection file

+++ Controllers/Lab2/App/lib/Conection.h
@@ -8,7 +8,7 @@ private:
 public:
 
 	Conection();
-	int receiveData();
+	float receiveData();
 	
 };
